add online user

+++ admin/includes/view_all_users.php
@@ -89,10 +89,13 @@ if (isset($_GET['change_to_sub'])) {
 
 // 使用者刪除
 if (isset($_GET['delete'])) {
-    $the_user_id = $_GET['delete'];
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
-    $delete_query = mysqli_query($connection, $query);
-    header("Location: users.php");
+    // 防止URL以及SQL injection
+    if (isset($_SESSION['user_role'])) {
+        $the_user_id =mysqli_real_escape_string($connection, $_GET['delete']);
+        $query = "DELETE FROM users WHERE user_id = {$the_user_id} ";
+        $delete_query = mysqli_query($connection, $query);
+        header("Location: users.php");
+    }
 }
 
 ?>
